Use dashboard modal layout
